drm/i915: Move intel_vblank_evade() & co. into intel_vblank.c
commit dea1731dfc25976e2046f71c03200aa3b642b34f upstream.

intel_vblank.c seems like the appropriate place for the core
vblank evasion code. Move it there.

deepin-Intel-SIG: commit dea1731dfc25 drm/i915: Move intel_vblank_evade() & co. into intel_vblank.c.
MTL Graphics backporting - 2024Q2

Link: https://patchwork.freedesktop.org/patch/msgid/20231213102519.13500-8-ville.syrjala@linux.intel.com
[ Quanxian Wang: amend commit log ]
